backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.0.1
torch==1.7.1
requests==2.31.0
python-dotenv==1.0.0
clip-by-openai==1.1 
pyyaml==6.0.2